# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TodoPay API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: parentId
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserList"
    post:
      operationId: createUser
      tags: [users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /users/{id}/reset-password:
    post:
      operationId: resetUserPassword
      tags: [users]
      summary: Reset user password
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password reset
  /users/{id}/permission:
    patch:
      operationId: updateUserPermission
      tags: [users]
      summary: Update user permission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserPermissionUpdate"
      responses:
        "200":
          description: Permission updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"

  # Members (회원)
...
  /virtual-accounts:
    get:
      operationId: listVirtualAccounts
      tags: [virtual-accounts]
      summary: List virtual accounts
      parameters:
        - name: memberId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of virtual accounts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VirtualAccountList"
  /virtual-accounts/{id}:
    get:
      operationId: getVirtualAccount
      tags: [virtual-accounts]
      summary: Get virtual account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Virtual account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VirtualAccount"

  # Withdrawals (출금 요청)
...
  /notices:
    get:
      operationId: listNotices
      tags: [notices]
      summary: List notices
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of notices
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NoticeList"
    post:
      operationId: createNotice
      tags: [notices]
      summary: Create notice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoticeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notice"
  /notices/{id}:
    get:
      operationId: getNotice
      tags: [notices]
      summary: Get notice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notice"
    patch:
      operationId: updateNotice
      tags: [notices]
      summary: Update notice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoticeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notice"
    delete:
      operationId: deleteNotice
      tags: [notices]
      summary: Delete notice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # OTP
...
components:
  schemas:
...
    AdminUser:
      type: object
      required: [id, loginId, name, role, permission, isActive, useOtp, createdAt]
      properties:
        id:
          type: integer
        loginId:
          type: string
        name:
          type: string
        role:
          type: string
          description: "superadmin | hq | distributor | agency | store"
        permission:
          type: string
          description: "readonly | admin | finance"
        isActive:
          type: boolean
        useOtp:
          type: boolean
        parentId:
          type: ["integer", "null"]
        parentName:
          type: ["string", "null"]
        createdAt:
          type: string
    AdminUserList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminUser"
        total:
          type: integer
    AdminUserInput:
      type: object
      required: [loginId, password, name, role, permission]
      properties:
        loginId:
          type: string
        password:
          type: string
        name:
          type: string
        role:
          type: string
        permission:
          type: string
        parentId:
          type: ["integer", "null"]
        useOtp:
          type: boolean
    AdminUserUpdate:
      type: object
      properties:
        name:
          type: string
        isActive:
          type: boolean
        permission:
          type: string
        useOtp:
          type: boolean
    ResetPasswordInput:
      type: object
      required: [newPassword]
      properties:
        newPassword:
          type: string
    UserPermissionUpdate:
      type: object
      required: [permission]
      properties:
        permission:
          type: string

    # Members
...
    VirtualAccount:
      type: object
      required: [id, accountNumber, bankName, status, memberId, memberName, createdAt]
      properties:
        id:
          type: integer
        accountNumber:
          type: string
        bankName:
          type: string
        status:
          type: string
          description: "active | revoked"
        memberId:
          type: integer
        memberName:
          type: string
        balance:
          type: number
        createdAt:
          type: string
    VirtualAccountList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/VirtualAccount"
        total:
          type: integer

    # Withdrawals
...
    Notice:
      type: object
      required: [id, title, content, isPinned, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        isPinned:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
    NoticeList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Notice"
        total:
          type: integer
    NoticeInput:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
        content:
          type: string
        isPinned:
          type: boolean
    NoticeUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        isPinned:
          type: boolean

    # OTP
...