# lib/api-zod/src/generated/types/adminUserInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TodoPay API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminUserInput {
  loginId: string;
  password: string;
  name: string;
  role: string;
  permission: string;
  /** @nullable */
  parentId?: number | null;
  useOtp?: boolean;
}